@@ -1,36 +1,40 @@
 <?php
 
 
-class IniFileLocatorTest extends PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+use Retrinko\Ini\Exceptions\FileException;
+use Retrinko\Ini\IniFileLocator;
+
+class IniFileLocatorTest extends TestCase
 {
 
     public function test_i_returnsIniFileLocatorInstance()
     {
-        $instance = \Retrinko\Ini\IniFileLocator::i();
-        $this->assertTrue($instance instanceof \Retrinko\Ini\IniFileLocator);
+        $instance = IniFileLocator::i();
+        $this->assertInstanceOf(IniFileLocator::class, $instance);
     }
 
     /**
      * @depends test_i_returnsIniFileLocatorInstance
-     * @throws \Retrinko\Ini\Exceptions\FileException
+     * @throws FileException
      */
     public function test_locate_returnsSameFileIfNoExistLocalFile()
     {
         $file = __DIR__.'/data/simple.ini';
-        $path = \Retrinko\Ini\IniFileLocator::i()->i()->locate($file);
+        $path = IniFileLocator::i()->i()->locate($file);
         $this->assertEquals(realpath($file), realpath($path));
     }
 
 
     /**
      * @depends test_i_returnsIniFileLocatorInstance
-     * @expectedException \Retrinko\Ini\Exceptions\FileException
-     * @throws \Retrinko\Ini\Exceptions\FileException
+     * @throws FileException
      */
     public function test_locate_unexistingFile_throwsException()
     {
         $file = __DIR__.'/data/nofile.ini';
-        \Retrinko\Ini\IniFileLocator::i()->i()->locate($file);
+        $this->expectException(FileException::class);
+        IniFileLocator::i()->i()->locate($file);
     }
 
 }
