@@ -1,40 +1,41 @@
-<?php
+<?php /** @noinspection PhpUnhandledExceptionInspection */
 
-class IniSectionTest extends PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+use Retrinko\Ini\Exceptions\InvalidDataException;
+use Retrinko\Ini\IniSection;
+
+class IniSectionTest extends TestCase
 {
 
     public function test_setContents_withProperParameters_returnsSection()
     {
         $contents = ['k1'=>'V1', 'k2'=>'V2', 'k3'=>'V3'];
-        $section = new \Retrinko\Ini\IniSection('section');
+        $section = new IniSection('section');
         $section = $section->setContents($contents);
-        $this->assertTrue($section instanceof \Retrinko\Ini\IniSection);
+        $this->assertInstanceOf(IniSection::class, $section);
     }
 
-    /**
-     * @expectedException \Retrinko\Ini\Exceptions\InvalidDataException
-     */
     public function test_setContents_withBadParameters_thrownsInvalidDataException()
     {
+        $this->expectException(InvalidDataException::class);
         $contents = 'bad contents';
-        $section = new \Retrinko\Ini\IniSection('section');
+        $section = new IniSection('section');
+        /** @noinspection PhpParamsInspection */
         $section->setContents($contents);
     }
 
     public function test_set_withProperParamaters_retrunsIniSection()
     {
-        $section = new \Retrinko\Ini\IniSection('test');
+        $section = new IniSection('test');
         $section = $section->set('hello', 'world');
-        $this->assertTrue($section instanceof \Retrinko\Ini\IniSection);
+        $this->assertInstanceOf(IniSection::class, $section);
     }
 
-    /**
-     * @expectedException \Retrinko\Ini\Exceptions\InvalidDataException
-     */
     public function test_set_withInvalidItemName_thrownsInvalidDataException()
     {
+        $this->expectException(InvalidDataException::class);
         $invalidItemName = 'hello!';
-        $section = new \Retrinko\Ini\IniSection('test');
+        $section = new IniSection('test');
         $section->set($invalidItemName, 'world');
     }
 
@@ -42,9 +43,9 @@ class IniSectionTest extends PHPUnit_Framework_TestCase
     {
         $parentContents = ['k0'=>'v0', 'k1'=>'v1', 'k2'=>'v2'];
         $contents = ['k1'=>'V1', 'k2'=>'V2', 'k3'=>'V3'];
-        $parent = new \Retrinko\Ini\IniSection('parentSection');
+        $parent = new IniSection('parentSection');
         $parent->setContents($parentContents);
-        $section = new \Retrinko\Ini\IniSection('test', $parent);
+        $section = new IniSection('test', $parent);
         $section->setContents($contents);
         $this->assertEquals('v0', $section->get('k0'));
         $this->assertEquals('V1', $section->get('k1'));
@@ -55,7 +56,7 @@ class IniSectionTest extends PHPUnit_Framework_TestCase
     public function test_get_withNotPresentFieldAndDefaultValue_returnsDefaultValue()
     {
         $contents = ['k1'=>'V1', 'k2'=>'V2', 'k3'=>'V3'];
-        $section = new \Retrinko\Ini\IniSection('test');
+        $section = new IniSection('test');
         $section->setContents($contents);
         $this->assertEquals('default', $section->get('noPresentField', 'default'));
     }
@@ -63,16 +64,16 @@ class IniSectionTest extends PHPUnit_Framework_TestCase
     public function test_hasField_returnsAlwaysBool()
     {
         $contents = ['k1'=>'V1', 'k2'=>'V2', 'k3'=>'V3'];
-        $section = new \Retrinko\Ini\IniSection('test');
+        $section = new IniSection('test');
         $section->setContents($contents);
-        $this->assertTrue(is_bool($section->hasItem('k2')));
-        $this->assertTrue(is_bool($section->hasItem('k0')));
+        $this->assertIsBool($section->hasItem('k2'));
+        $this->assertIsBool($section->hasItem('k0'));
     }
 
     public function test_hasField_withPresentField_returnsTrue()
     {
         $contents = ['k1'=>'V1', 'k2'=>'V2', 'k3'=>'V3'];
-        $section = new \Retrinko\Ini\IniSection('test');
+        $section = new IniSection('test');
         $section->setContents($contents);
         $this->assertTrue($section->hasItem('k1'));
 
@@ -81,7 +82,7 @@ class IniSectionTest extends PHPUnit_Framework_TestCase
     public function test_hasField_withNotPresentField_returnsFalse()
     {
         $contents = ['k1'=>'V1', 'k2'=>'V2', 'k3'=>'V3'];
-        $section = new \Retrinko\Ini\IniSection('test');
+        $section = new IniSection('test');
         $section->setContents($contents);
         $this->assertFalse($section->hasItem('k0'));
     }
@@ -90,9 +91,9 @@ class IniSectionTest extends PHPUnit_Framework_TestCase
     {
         $parentContents = ['k0'=>'v0', 'k1'=>'v1'];
         $contents = ['k1'=>'V1', 'k2'=>'v2'];
-        $parent = new \Retrinko\Ini\IniSection('parentSection');
+        $parent = new IniSection('parentSection');
         $parent->setContents($parentContents);
-        $section = new \Retrinko\Ini\IniSection('test', $parent);
+        $section = new IniSection('test', $parent);
         $section->setContents($contents);
 
         $this->assertEquals(['k0'=>'v0', 'k1'=>'V1', 'k2'=>'v2'], $section->toArray());
@@ -100,7 +101,7 @@ class IniSectionTest extends PHPUnit_Framework_TestCase
 
     public function test_toString_returnsProperString()
     {
-        $section = new \Retrinko\Ini\IniSection('test');
+        $section = new IniSection('test');
         $section->set('hello', 'world');
         $section->set('bool', true);
         $section->set('int', 3);
